Happy accident
I stubbled upon a nice dance as a result of joining two animations together. I plan to reverse engineer the animation.

So far I have implemented the idle 3 animation but need more work on the timing.

I separated the lips and now joe can smile or frown when needed in any animation