Start
Asks the GiftBatchHandler for gifts by calling TakeRandomGifts(). Based on the gifts it gets, it animates Joe reaction to that gift. Remember It will get a list of gifts from the function, so it will go index by index and react to each gift. How fast it reacts to all the gifts will depend on the number of gifts the TakeRandomGifts() returns.