Joes_Animation_Manager Done
Now scripts pull gift from TikTok live and uses that to animate Joe.

Fixed Bugs where a animation function was not called. This happened because we were trying to call a function that handles UI(animation) logic in a background thread. When we use the web socket(on("message" ...), it calls everything within it in a background thread.