Testsed The new logic with actual ticktok live
I started having issues were my code wouldn't run because my code ran processes that are required to run on the main thread, on the background thread . The problem originates from the onMessage() in the TikTokListener. It calls addToGiftIdContainer() in the background thread so whatever process underneath  that function runs on the background thread.