Now Joe can activate multiple idel state when no  gifts are coming in
- I fixed a concurrency issue.

Description: When GetARandomGift is called and finds the gift bag empty, it triggers CollectMoreGifts, which in turn may fire the GetNextGiftEvent event before GetARandomGift returns null. Leading to playNextAnimation2 being called (through the event listener) while another instance of playNextAnimation might still be active, leading to concurrency issues with the lock-protected variables (InPlayMode, InIdleState).

Fix: Added a delay to CollectMoreGifts so that if it fires and event, the event will be fired slightly late giving any instance of playNextAnimation() chance to finish running before another instance is created