Made a Big banner of clouds and moved them
I thought of using a spawnner system to spawn the clouds when needed but I realized that it may be too complicated to do. Instead I will use three large banners. One for the big clouds, another for the mideum sized clouds and last one for the small sized clouds